+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The API consists of all public Java types from `com.atlassian.performance.tools.
 [1.1.0]: https://bitbucket.org/atlassian/infrastructure/branches/compare/release-1.1.0%0Drelease-1.0.0
 ### Added 
 - Add diagnosticsLimit parameter to limit how many times diagnostics can be executed.
+- Add maxConnections parameter to override MySQL max_connections value.
 - Explain [contribution guidelines](CONTRIBUTING.md). 
 
 ## [1.0.0] - 2018-08-24